Refactor config-apollo component, use config-center ways
